+++ bevy_mod_scripting_core/src/asset.rs
@@ -1,0 +1,8 @@
+use bevy::asset::Asset;
+
+/// All code assets share this common interface.
+/// When adding a new code asset don't forget to implement asset loading
+/// and inserting appropriate systems when registering with the app
+pub trait CodeAsset: Asset {
+    fn bytes(&self) -> &[u8];
+}
