+++ bevy_mod_scripting_core/src/docs.rs
@@ -1,4 +1,5 @@
-use crate::ScriptError;
+use crate::error::ScriptError;
+
 
 /// A documentation piece exported by an `APIProvider`
 pub trait DocFragment: 'static {
