+++ languages/bevy_mod_scripting_lua/src/docs.rs
@@ -8,7 +8,7 @@ use std::{
 use bevy::asset::FileAssetIo;
 use tealr::TypeWalker;
 
-use bevy_mod_scripting_core::{DocFragment, ScriptError};
+use bevy_mod_scripting_core::prelude::*;
 
 pub type TypeWalkerBuilder = fn(TypeWalker) -> TypeWalker;
