+++ languages/bevy_mod_scripting_rhai/src/lib.rs
@@ -4,8 +4,8 @@ use crate::{
 };
 use bevy::prelude::*;
 use bevy_mod_scripting_core::{
-    prelude::*, script_add_synchronizer, script_hot_reload_handler, script_remove_synchronizer,
-    APIProviders, CachedScriptEventState,
+    prelude::*, 
+    systems::*
 };
 use rhai::*;
 use std::marker::PhantomData;
